qwen: fail on OOM, and stop accumulating window scores in the KV dtype
Three fixes, all aimed at the same thing: the Qwen2.5 LongBench column was not
measuring what the Llama-3.1 and Mistral-v0.2 columns measure.

1. skip_oom: false, resume: false on both Qwen "ours" configs.

   They were the only "ours" configs in the repo with either set true.
   skip_oom: true records an OOM'd example as pred=null, and the scorer counts a
   null as 0 while keeping it in the denominator (correctly — dropping it would
   inflate the mean). So an OOM did not fail the run: it returned a complete,
   plausible, uniformly depressed score table with `skipped=N` in the log as the
   only tell. resume: true then skipped any dataset whose jsonl had > 0 lines, so
   a file truncated by a crashed run was accepted as finished and scored over
   however many examples happened to be written. Two independent ways for this
   column alone to report a number that is not a method result.

   longbench.max_length is null here (YaRN 131072, so nothing truncates), which
   is what puts narrativeqa's 30k+ prompts through the auxiliary score pass in
   the first place.

2. The window-score accumulators are fp32, in both hooks.

   state.window_scores is a running sum over the T query rows of a pass and then
   over every decode step, and its dtype came from whatever the hook allocated —
   q.dtype in the flash hook, attn_weights.dtype in the eager one. In a
   reduced-precision accumulator an addend below half an ULP of the running total
   rounds to no change and is discarded outright. Measured on a realistic
   trajectory: ~76% of per-decode-step contributions vanish in bfloat16, ~52% in
   float16, and the retained top-k set diverges from the exact answer by ~9% on a
   512-step generation (gov_report, the longest max_gen in the suite).

   Qwen2.5 is the only bf16 column — 8 mantissa bits against float16's 11 — so it
   lost twice what the other two did and its rankings were not built the same way.
   The fix moves all three onto an exact sum rather than onto the same rounding,
   so Llama and Mistral are NOT byte-identical to before and need re-running if
   their numbers are to sit in the same table.

   Cost is [B, H_q, W] in fp32: ~258 KB/layer at Qwen's 28 heads. The softmax
   intermediate is untouched and stays on _score_softmax_dtype() — torch reduces
   bf16/fp16 softmax with an fp32 internal accumulator, so its denominator was
   never the lossy part. In the flash hook the widening is applied to the reduced
   [B,H_kv,rep,S] per-chunk result, not to the [..,blk,S] block, so the largest
   transient in the prefill pass does not double.

3. The eager-attention ceiling has a num_layers factor.

   The eager backend needs output_attentions=True to score at all, and
   transformers does not hand that back one layer at a time: <Model>Model.forward
   accumulates every layer's attn_weights into all_self_attns and returns the
   whole tuple, so all num_hidden_layers [B, H_q, T, T] matrices are resident at
   once. The guard reported ONE layer's worth, understating the requirement by 28x
   on Qwen2.5-7B: it put its 8 GB warning threshold at ~11.9k tokens when the real
   one is ~2.3k, i.e. it stayed quiet through the whole range LongBench lives in.
   The suggested max_length was likewise sqrt(num_layers) too long. Both now
   account for every layer, and the message says to set skip_oom: false.

Tests: tests/test_score_precision.py (21). The arithmetic and guard tests run
anywhere — the three substantive guard assertions fail against the pre-fix guard,
confirmed. The 12 model-driven dtype tests are gated on
is_transformers_version_supported: past 4.47.1 the Cache ABI changed and even a
bare model.forward with this cache raises from the mask builder, so they are
skipped rather than reported as anything. Run them in the pinned 4.47.1 env.

Full suite: 84 failures before and after, byte-identical lists — all from the
unpinned transformers on this box, none added.

+++ modules/windowed_eager_cache/hooks.py
@@ -172,7 +172,17 @@ def install_score_hooks(
                 # attn_weights: [B, H_q, T, S] over the effective-K axis. Sum
                 # across the T (query) axis — every query row contributes (H2O
                 # cumulative, no obs_window) — to per-key received attention.
-                token_scores = attn_weights.sum(dim=-2)          # [B, H_q, S]
+                #
+                # fp32, mirroring modules/windowed_cache/hooks.py. This sum itself
+                # is fine in either dtype (torch reduces bf16/fp16 with an fp32
+                # internal accumulator), but its dtype is inherited by
+                # `state.window_scores`, which is then accumulated across every
+                # decode step by `scorer.accumulate`. That is the lossy one: in a
+                # reduced-precision accumulator a per-step contribution below half
+                # an ULP of the running total is discarded outright — 76% of them
+                # in bfloat16, 52% in float16. The window ranking is the method, so
+                # the running total is kept exact.
+                token_scores = attn_weights.sum(dim=-2, dtype=torch.float32)  # [B,H_q,S]
 
                 # At q > 0 the effective K is the unsorted [sink ‖ body ‖ Q]
                 # layout, so the key axis of attn_weights is unsorted too; undo
